requirements.txt
ollama
langchain
langchain-community
langchain-openai
faiss-cpu
pypdf
python-dotenv
langchain-ollama